+++ lib/mix/tasks/brando.gen.html.ex
@@ -30,18 +30,26 @@ defmodule Mix.Tasks.Brando.Gen.Html do
 
     attrs        = Mix.Brando.attrs(attrs)
     villain?     = :villain in Dict.values(attrs)
+    sequenced?   = Mix.shell.yes?("\nMake schema sequenceable?")
     image_field? = :image in Dict.values(attrs)
     binding      = Mix.Brando.inflect(singular)
     admin_path   = Enum.join(["admin", binding[:path]], "_")
     path         = binding[:path]
     route        = path
-                   |> String.split("/") |> Enum.drop(-1)
-                   |> Kernel.++([plural]) |> Enum.join("/")
+                   |> String.split("/")
+                   |> Enum.drop(-1)
+                   |> Kernel.++([plural])
+                   |> Enum.join("/")
     admin_module = Enum.join([binding[:base], "Admin", binding[:scoped]], ".")
-    binding      = binding ++ [plural: plural, route: route,
-                               image_field: image_field?, villain: villain?,
-                               admin_module: admin_module, admin_path: admin_path,
-                               inputs: inputs(attrs), params: Mix.Brando.params(attrs)]
+    binding      = binding ++ [plural: plural,
+                               route: route,
+                               image_field: image_field?,
+                               villain: villain?,
+                               sequenced: sequenced?,
+                               admin_module: admin_module,
+                               admin_path: admin_path,
+                               inputs: inputs(attrs),
+                               params: Mix.Brando.params(attrs)]
 
     files = [
       {:eex, "admin_controller.ex",
@@ -85,6 +93,14 @@ defmodule Mix.Tasks.Brando.Gen.Html do
       ]
     end
 
+    if sequenced? do
+      files = files ++ [
+        {:eex, "sequence.html.eex",
+               "web/templates/admin/#{path}/sequence.html.eex"}
+      ]
+      args = args ++ ["--sequenced"]
+    end
+
     Mix.Brando.check_module_name_availability!(binding[:module] <> "Controller")
     Mix.Brando.check_module_name_availability!(binding[:module] <> "View")
     Mix.Brando.check_module_name_availability!(binding[:admin_module] <> "Controller")
@@ -98,6 +114,16 @@ defmodule Mix.Tasks.Brando.Gen.Html do
     villain_info =
       if villain? do
         ~s(    villain_routes "/#{route}",    #{binding[:scoped]}Controller)
+      else
+        ""
+      end
+
+    sequenced_info =
+      if sequenced? do
+        """
+            get    "/#{route}/sort", #{binding[:scoped]}Controller, :sequence
+            post   "/#{route}/sort", #{binding[:scoped]}Controller, :sequence_post
+        """
       else
         ""
       end
@@ -121,7 +147,7 @@ defmodule Mix.Tasks.Brando.Gen.Html do
         patch  "/#{route}/:id",        #{binding[:scoped]}Controller, :update
         put    "/#{route}/:id",        #{binding[:scoped]}Controller, :update
 
-    """ <> villain_info <>
+    """ <> villain_info <> sequenced_info <>
     """
 
     and then update your repository by running migrations: